+++ app/src/main/java/woowacourse/shopping/model/uimodel/ProductUIModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.uimodel
+package woowacourse.shopping.model.uimodel
 
 import java.io.Serializable
