+++ app/src/main/java/woowacourse/shopping/model/uimodel/RecentProductUIModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.uimodel
+package woowacourse.shopping.model.uimodel
 
 class RecentProductUIModel(
     val productUIModel: ProductUIModel
