+++ app/src/main/java/woowacourse/shopping/model/uimodel/mapper/ProductMapper.kt
@@ -1,11 +1,11 @@
-package woowacourse.shopping.uimodel.mapper
+package woowacourse.shopping.model.uimodel.mapper
 
 import com.shopping.domain.CartProduct
 import com.shopping.domain.Product
 import com.shopping.domain.RecentProduct
-import woowacourse.shopping.uimodel.CartProductUIModel
-import woowacourse.shopping.uimodel.ProductUIModel
-import woowacourse.shopping.uimodel.RecentProductUIModel
+import woowacourse.shopping.model.uimodel.CartProductUIModel
+import woowacourse.shopping.model.uimodel.ProductUIModel
+import woowacourse.shopping.model.uimodel.RecentProductUIModel
 
 fun ProductUIModel.toDomain() =
     Product(id, name, url, price)
