+++ app/src/main/res/layout/item_cart_product.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="cartProduct"
-            type="woowacourse.shopping.uimodel.CartProductUIModel" />
+            type="woowacourse.shopping.model.uimodel.CartProductUIModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
