+++ app/src/main/res/layout/item_product_main.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="product"
-            type="woowacourse.shopping.uimodel.ProductUIModel" />
+            type="woowacourse.shopping.model.uimodel.ProductUIModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
